+++ coloraide/gamut/fit_raytrace.py
@@ -245,29 +245,17 @@ class RayTrace(Fit):
 
             # Create a line from our color to color with zero lightness.
             # Trace the line to the RGB cube finding the face and the point where it intersects.
+            # Take two rounds to get us as close as we can get.
             size = [1.0, 1.0, 1.0]
-            face, intersection = raytrace_cube(size, gamutcolor[:-1], achroma)
-            if face:
-                gamutcolor[:-1] = intersection
-                gamutcolor.set(
-                    {
-                        L: mapcolor[l],
-                        H: mapcolor[h]
-                    }
-                )
-
-                # See if we were moved back outside gamut after correcting lightness and hue
+            for _ in range(2):
                 face, intersection = raytrace_cube(size, gamutcolor[:-1], achroma)
                 if face:
                     gamutcolor[:-1] = intersection
-                    gamutcolor.set(
-                        {
-                            L: mapcolor[l],
-                            H: mapcolor[h]
-                        }
-                    )
+                    gamutcolor.set({L: mapcolor[l], H: mapcolor[h]})
 
-            # We be under saturated now, so extended the vector out and find the surface one last time
+            # We might be under saturated now, so extend the vector out,
+            # ignoring the original point and find the surface one last
+            # Give us the most saturated color at that point on the RGB cube.
             x1, y1, z1 = achroma
             x2, y2, z2 = gamutcolor[:-1]
             x3 = x2 + (x2 - x1) * 100
@@ -276,7 +264,9 @@ class RayTrace(Fit):
             face, intersection = raytrace_cube(size, [x3, y3, z3], achroma)
             if face:
                 gamutcolor[:-1] = intersection
+                gamutcolor.set({L: mapcolor[l], H: mapcolor[h]})
 
+            gamutcolor[:-1] = [alg.clamp(x, 0, 1) for x in gamutcolor[:-1]]
             color.update(gamutcolor)
 
         if coerced: