+++ tests/test_a98_rgb.py
@@ -54,7 +54,7 @@ class TestA98RGBSerialize(util.ColorAssertsPyTest):
         (
             'color(a98-rgb 1.2 0.2 0)',
             {'color': True, 'fit': 'lch-raytrace'},
-            'color(a98-rgb 1 0.53149 0.37343)'
+            'color(a98-rgb 1 0.52561 0.38454)'
         ),
         ('color(a98-rgb 1.2 0.2 0)', {'fit': False}, 'color(a98-rgb 1.2 0.2 0)')
     ]